+++ src/KurzSharp.EmbeddedSource/Hooks/IBeforeCreateHook.cs
@@ -1,6 +1,0 @@
-namespace KurzSharp;
-
-public interface IBeforeCreateHook
-{
-    void OnBeforeCreate();
-}