+++ src/KurzSharp.EmbeddedSource/Hooks/IBeforeDeleteHook.cs
@@ -1,6 +1,0 @@
-namespace KurzSharp;
-
-public interface IBeforeDeleteHook
-{
-    void OnBeforeDelete();
-}