+++ src/KurzSharp.EmbeddedSource/Hooks/IBeforeReadHook.cs
@@ -1,6 +1,0 @@
-namespace KurzSharp;
-
-public interface IBeforeReadHook
-{
-    void OnBeforeRead();
-}