+++ src/KurzSharp.EmbeddedSource/Hooks/IBeforeUpdateHook.cs
@@ -1,6 +1,0 @@
-namespace KurzSharp;
-
-public interface IBeforeUpdateHook
-{
-    void OnBeforeUpdate();
-}