+++ src/KurzSharp.Templates/Models/PlaceholderModel.cs
@@ -1,6 +1,20 @@
-namespace KurzSharp;
+#if NET5_0_OR_GREATER
+using System.Text.Json.Serialization;
+using KurzSharp;
 
-public class PlaceholderModel
+namespace KurzSharp.Templates;
+
+public partial class PlaceholderModel : LifecycleHooks<PlaceholderModelDto>
 {
+    [JsonConstructor]
+    private PlaceholderModel()
+    {
+    }
+
     public Guid Id { get; set; }
+}
+#endif
+
+public partial class PlaceholderModel
+{
 }