+++ src/KurzSharp/RestApiSourceGenerator.cs
@@ -1,10 +1,12 @@
 using System.Collections.Immutable;
-using System.Diagnostics;
 using System.Text;
+using KurzSharp.Templates;
 using KurzSharp.Utils;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Text;
+using KurzSharp.Templates.RestApi;
+using KurzSharp.Templates.Database;
 
 namespace KurzSharp;
 
@@ -12,13 +14,6 @@ namespace KurzSharp;
 public class RestApiSourceGenerator : IIncrementalGenerator
 {
     private const string TemplatesNamespace = "KurzSharp.Templates";
-    private const string RestApiTemplatesNamespace = $"{TemplatesNamespace}.RestApi";
-    private const string DbTemplatesNamespace = $"{TemplatesNamespace}.Database";
-
-    private const string PlaceholderTypeName = nameof(PlaceholderModel);
-
-    // NOTE: Can't use `nameof` as that would error on NET7 code being in .netstandard 2.0 (Source Gen) Code
-    private const string DbContextName = "KurzSharpDbContext";
 
     public void Initialize(IncrementalGeneratorInitializationContext context)
     {
@@ -35,71 +30,118 @@ public class RestApiSourceGenerator : IIncrementalGenerator
     {
         var (compilation, classDeclarationsWithAttrs) = arg2;
 
-        var typesAndNamespaces = new List<(string type, string typeNamespace)>();
+        var typesAndNamespaces = new List<ModelSourceInfo>();
+        var namedTypeSymbols = new List<INamedTypeSymbol>();
 
         foreach (var (syntax, _) in classDeclarationsWithAttrs)
         {
             var symbol = compilation.GetSemanticModel(syntax.SyntaxTree).GetDeclaredSymbol(syntax);
 
-            if (symbol is null)
+            if (symbol is not INamedTypeSymbol typeSymbol)
             {
                 continue;
             }
 
+            namedTypeSymbols.Add(typeSymbol);
             var typeName = syntax.Identifier.ToString();
             var typeNamespace = symbol.ContainingNamespace.ToDisplayString();
 
-            typesAndNamespaces.Add((typeName, typeNamespace));
+            typesAndNamespaces.Add(new ModelSourceInfo(typeName, typeNamespace, typeSymbol));
 
-            AddController(typeName, typeNamespace, syntax, ctx);
+            AddModelDto(typeName, typeNamespace, typeSymbol, ctx);
+            AddPartialModel(typeName, typeNamespace, ctx);
+            AddController(typeName, typeNamespace, ctx);
         }
 
         AddDbContext(typesAndNamespaces, ctx);
-        AddSetupExtension(typesAndNamespaces, ctx);
+        AddSetupExtension(typesAndNamespaces, namedTypeSymbols, ctx);
     }
 
-    private static void AddController(string typeName, string typeNamespace, ClassDeclarationSyntax syntax,
-        SourceProductionContext ctx)
+    private static void AddController(string typeName, string typeNamespace, SourceProductionContext ctx)
     {
-        var controllerName = $"{typeName}Controller";
-        // NOTE: Can't use `nameof` as that would error on NET7 code being in .netstandard 2.0 (Source Gen) Code
-        const string placeholderControllerName = $"{PlaceholderTypeName}Controller";
-
         var controllerSource =
-            TemplatesUtils.GetTemplateFileContent(RestApiTemplatesNamespace, placeholderControllerName);
+            TemplatesUtils.GetTemplateFileContent($"{TemplatesNamespace}.RestApi", nameof(PlaceholderModelController));
 
-        var source = controllerSource.InjectHooks(syntax).ReplacePlaceholderType(typeName).FixupNamespaces()
+        var source = controllerSource.FixupNamespaces()
+            .ReplacePlaceholderType(typeName)
             .AddUsing(typeNamespace);
 
-        ctx.AddSource($"{controllerName}.g.cs", SourceText.From(source, Encoding.UTF8));
+        ctx.AddSource($"{typeName}Controller.g.cs", SourceText.From(source, Encoding.UTF8));
     }
 
-    private static void AddDbContext(IList<(string type, string typeNamespace)> typesAndNamespaces,
+    private static void AddModelDto(string typeName, string typeNamespace, INamespaceOrTypeSymbol typeSymbol,
         SourceProductionContext ctx)
     {
-        var typeNames = typesAndNamespaces.Select(t => t.type);
-        var typeNamespaces = typesAndNamespaces.Select(t => t.typeNamespace);
+        var props = typeSymbol.GetMembers()
+            .Where(s => s.Kind == SymbolKind.Property)
+            .Cast<IPropertySymbol>()
+            .ToArray();
 
-        var dbContextSource =
-            TemplatesUtils.GetTemplateFileContent(DbTemplatesNamespace, DbContextName);
+        var propsDeclarations = string.Join("\n",
+            props.Select(s => $"{s.DeclaredAccessibility.ToString().ToLower()} {s.Type} {s.Name} {{ get; set; }}"));
 
-        const string placeholderDbSet = $"public DbSet<{PlaceholderTypeName}> {PlaceholderTypeName}s {{ get; set; }}";
+        var propsMaps = string.Join("\n", props.Select(p => $"model.{p.Name} = {p.Name};"));
 
-        var dbSets = string.Join("\n", typeNames.Select(t => placeholderDbSet.ReplacePlaceholderType(t)));
+        var source = TemplatesUtils
+            .GetTemplateFileContent($"{TemplatesNamespace}.Models", $"{TemplatesUtils.PlaceholderTypeName}Dto")
+            .Replace(TemplatesNamespace, typeNamespace)
+            .Replace("public Guid PlaceholderId { get; set; }", propsDeclarations)
+            .Replace("model.Id = PlaceholderId;", propsMaps)
+            .ReplacePlaceholderType(typeName)
+            .AddUsing(typeNamespace);
 
-        var source = dbContextSource.Replace(placeholderDbSet, dbSets)
-            .FixupNamespaces().AddUsing(typeNamespaces);
+        var fileName = $"{TemplatesUtils.PlaceholderTypeName.ReplacePlaceholderType(typeName)}Dto";
 
-        ctx.AddSource($"{DbContextName}.g.cs", SourceText.From(source, Encoding.UTF8));
+        ctx.AddSource($"{fileName}.g.cs", SourceText.From(source, Encoding.UTF8));
     }
 
-    private static void AddSetupExtension(
-        IEnumerable<(string type, string typeNamespace)> typesAndNamespaces, SourceProductionContext ctx)
+    private static void AddPartialModel(string typeName, string typeNamespace, SourceProductionContext ctx)
     {
-        const string extensionSetupFileName = "KurzSharpSetupExtension";
+        var source = TemplatesUtils
+            .GetTemplateFileContent($"{TemplatesNamespace}.Models", TemplatesUtils.PlaceholderTypeName)
+            .Replace("public Guid Id { get; set; }", string.Empty)
+            .Replace(TemplatesNamespace, typeNamespace)
+            .ReplacePlaceholderType(typeName)
+            .AddUsing(typeNamespace);
+
+        var fileName = TemplatesUtils.PlaceholderTypeName.ReplacePlaceholderType(typeName);
+
+        ctx.AddSource($"{fileName}.g.cs", SourceText.From(source, Encoding.UTF8));
+    }
+
+    private static void AddDbContext(IList<ModelSourceInfo> modelSourceInfos, SourceProductionContext ctx)
+    {
+        const string dbTemplatesNamespace = $"{TemplatesNamespace}.Database";
+        const string dbContextName = nameof(KurzSharpDbContext);
+
+        var dbContextSource = TemplatesUtils.GetTemplateFileContent(dbTemplatesNamespace, dbContextName);
+
+        const string placeholderDbSet =
+            $"public DbSet<{TemplatesUtils.PlaceholderDtoTypeName}> {TemplatesUtils.PlaceholderTypeName}s {{ get; set; }}";
+
+        var dbSets = string.Join("\n",
+            modelSourceInfos.Select(t => placeholderDbSet.ReplacePlaceholderType(t.TypeName)));
+
+        var source = dbContextSource.Replace(placeholderDbSet, dbSets)
+            .FixupNamespaces()
+            .AddUsing(modelSourceInfos.Select(t => t.TypeNamespace));
+
+        ctx.AddSource($"{dbContextName}.g.cs", SourceText.From(source, Encoding.UTF8));
+    }
+
+    private static void AddSetupExtension(IEnumerable<ModelSourceInfo> modelSourceInfos, List<INamedTypeSymbol> syntax,
+        SourceProductionContext ctx)
+    {
+        const string extensionSetupFileName = nameof(KurzSharpSetupExtension);
+
         var setupExtSource = TemplatesUtils.GetTemplateFileContent(TemplatesNamespace, extensionSetupFileName);
 
-        var source = setupExtSource.FixupNamespaces().AddUsing(typesAndNamespaces.Select(t => t.typeNamespace));
+        var registeredTypes = syntax.Aggregate(string.Empty,
+            (acc, typeSymbol) => acc + $"services.AddSingleton<{typeSymbol.Name}>();");
+
+        var source = setupExtSource.Replace("services.AddTransient<PlaceholderModel>();", registeredTypes)
+            .FixupNamespaces()
+            .AddUsing(modelSourceInfos.Select(t => t.TypeNamespace));
 
         ctx.AddSource($"{extensionSetupFileName}.g.cs", SourceText.From(source, Encoding.UTF8));
     }