+++ src/KurzSharp/Utils/DeclarationsUtils.cs
@@ -12,7 +12,7 @@ public static class DeclarationsUtils
         where T : TypeDeclarationSyntax
     {
         var attrName = typeof(TAttribute).Name;
-        var attrNameShort = attrName.Replace(AttributeOptionalPostfix, "");
+        var attrNameShort = attrName.Replace(AttributeOptionalPostfix, string.Empty);
 
         return context.SyntaxProvider.CreateSyntaxProvider((node, _) =>
                 node is T
@@ -30,10 +30,5 @@ public static class DeclarationsUtils
 
                 return (syntax, attribute);
             });
-    }
-    
-    public static bool HasInterface(this ClassDeclarationSyntax source, string interfaceName)
-    {
-        return source.BaseList?.Types.Any(baseType => baseType.ToString().Contains(interfaceName)) ?? false;
     }
 }